fix bugs on routes
