+++ src/Components/Display.js
@@ -1,4 +1,5 @@
 import Movie from "./Movie";
+import { useEffect } from "react";
 
 function Display({
   setModalIsOpen,
@@ -6,7 +7,36 @@ function Display({
   selectedMovieID,
   setSelectedMovieID,
   apiMovieData,
+  setError,
+  setApiMovieData,
 }) {
+  function getMoviesFromApi() {
+    fetch("https://rancid-tomatillos.herokuapp.com/api/v2/movies")
+      .then((res) => {
+        if (!res.ok) {
+          throw new Error(
+            "Oops! Something went wrong on the server. Please try again later."
+          );
+        } else {
+          return res.json();
+        }
+      })
+      .then((data) => {
+        console.log(data);
+        setApiMovieData(data.movies);
+      })
+      .catch((error) => {
+        setError(error.message || "An unknown error occurred.");
+        console.log(error);
+      });
+  }
+
+  useEffect(() => {
+    getMoviesFromApi();
+    setApiMovieData(apiMovieData);
+    setModalIsOpen(false);
+  }, []);
+
   return (
     <div
       style={{
