+++ src/Components/Movie.js
@@ -1,26 +1,13 @@
 import StarRating from "./StarRating";
 import "./Movie.scss";
 import { useNavigate } from "react-router-dom";
-function Movie({
-  movieObj,
-  setModalIsOpen,
-  modalIsOpen,
-  setSelectedMovieID,
-}) {
+function Movie({ movieObj, setModalIsOpen, modalIsOpen, setSelectedMovieID }) {
   const navigate = useNavigate();
-  const toggleModal = () => {
-    setModalIsOpen(!modalIsOpen);
-    if (!modalIsOpen) {
-      localStorage.setItem('modalState', 'open');
-    } else {
-      localStorage.removeItem('modalState');
-    }
-  };
 
   function handleMovieClick(e) {
-    toggleModal();
-    setSelectedMovieID(parseInt(e.target.id));
-    navigate(`/${parseInt(e.target.id)}`);
+    setModalIsOpen(!modalIsOpen);
+    setSelectedMovieID(movieObj.id);
+    navigate(`/${movieObj.id}`);
   }
   return (
     <div className="movie" id={movieObj.id}>
